[DEBT-043-D13] Split analysis/liquidity.py into focused modules
- Split 461 LoC file into 5 focused modules:
  - _liquidity_models.py (87): Data classes and enums
  - _depth.py (67): Orderbook depth scoring
  - _slippage.py (170): Slippage estimation and constraints
  - _scoring.py (149): Composite liquidity scoring
  - liquidity.py (94): Re-exports + OrderbookAnalyzer class
- All files well under 400 line ceiling
- Public API preserved via re-exports in liquidity.py
- Quality gates: pre-commit, pytest (1052 passed)
